requirements.txt
python-dateutil==2.9.0.post0
pygame==2.6.1
pymunk==6.11.1
requests==2.32.3